adapt tests, remove SyncApi and only use API
