--- docs/AsyncApi.js.html
@@ -1,111 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    
-    <meta charset="utf-8">
-    <title>AsyncApi.js - Documentation</title>
-    
-    
-    <script src="scripts/prettify/prettify.js"></script>
-    <script src="scripts/prettify/lang-css.js"></script>
-    <!--[if lt IE 9]>
-      <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
-    <![endif]-->
-    <link type="text/css" rel="stylesheet" href="styles/prettify.css">
-    <link type="text/css" rel="stylesheet" href="styles/jsdoc.css">
-    <script src="scripts/nav.js" defer></script>
-    
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-</head>
-<body>
-
-<input type="checkbox" id="nav-trigger" class="nav-trigger" />
-<label for="nav-trigger" class="navicon-button x">
-  <div class="navicon"></div>
-</label>
-
-<label for="nav-trigger" class="overlay"></label>
-
-<nav >
-    
-    
-    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="Api.html">Api</a></li><li><a href="AsyncApi.html">AsyncApi</a><ul class='methods'><li data-type='method'><a href="AsyncApi.html#doCall">doCall</a></li></ul></li><li><a href="ColumnBuilder.html">ColumnBuilder</a><ul class='methods'><li data-type='method'><a href="ColumnBuilder.html#addConstraint">addConstraint</a></li><li data-type='method'><a href="ColumnBuilder.html#decimalDigits">decimalDigits</a></li><li data-type='method'><a href="ColumnBuilder.html#description">description</a></li><li data-type='method'><a href="ColumnBuilder.html#displayName">displayName</a></li><li data-type='method'><a href="ColumnBuilder.html#formatPattern">formatPattern</a></li><li data-type='method'><a href="ColumnBuilder.html#groups">groups</a></li><li data-type='method'><a href="ColumnBuilder.html#hidden">hidden</a></li><li data-type='method'><a href="ColumnBuilder.html#identifier">identifier</a></li><li data-type='method'><a href="ColumnBuilder.html#languageType">languageType</a></li><li data-type='method'><a href="ColumnBuilder.html#link">link</a></li><li data-type='method'><a href="ColumnBuilder.html#maxLength">maxLength</a></li><li data-type='method'><a href="ColumnBuilder.html#minLength">minLength</a></li><li data-type='method'><a href="ColumnBuilder.html#multilanguage">multilanguage</a></li><li data-type='method'><a href="ColumnBuilder.html#ordering">ordering</a></li><li data-type='method'><a href="ColumnBuilder.html#separator">separator</a></li><li data-type='method'><a href="ColumnBuilder.html#showMemberColumns">showMemberColumns</a></li><li data-type='method'><a href="ColumnBuilder.html#simpleLink">simpleLink</a></li><li data-type='method'><a href="ColumnBuilder.html#singleDirection">singleDirection</a></li><li data-type='method'><a href="ColumnBuilder.html#toDescription">toDescription</a></li><li data-type='method'><a href="ColumnBuilder.html#toDisplayName">toDisplayName</a></li><li data-type='method'><a href="ColumnBuilder.html#toName">toName</a></li><li data-type='method'><a href="ColumnBuilder.html#toOrdering">toOrdering</a></li></ul></li><li><a href="ConstraintBuilder.html">ConstraintBuilder</a><ul class='methods'><li data-type='method'><a href="ConstraintBuilder.html#.archiveCascade">archiveCascade</a></li><li data-type='method'><a href="ConstraintBuilder.html#.cardinality">cardinality</a></li><li data-type='method'><a href="ConstraintBuilder.html#.cardinalityManyToOne">cardinalityManyToOne</a></li><li data-type='method'><a href="ConstraintBuilder.html#.cardinalityOneToMany">cardinalityOneToMany</a></li><li data-type='method'><a href="ConstraintBuilder.html#.cardinalityOneToOne">cardinalityOneToOne</a></li><li data-type='method'><a href="ConstraintBuilder.html#.deleteCascade">deleteCascade</a></li><li data-type='method'><a href="ConstraintBuilder.html#.finalCascade">finalCascade</a></li></ul></li><li><a href="SyncApi.html">SyncApi</a><ul class='methods'><li data-type='method'><a href="SyncApi.html#createColumn">createColumn</a></li><li data-type='method'><a href="SyncApi.html#createColumns">createColumns</a></li><li data-type='method'><a href="SyncApi.html#createRow">createRow</a></li><li data-type='method'><a href="SyncApi.html#createRows">createRows</a></li><li data-type='method'><a href="SyncApi.html#createTable">createTable</a></li><li data-type='method'><a href="SyncApi.html#doCall">doCall</a></li><li data-type='method'><a href="SyncApi.html#fetchTable">fetchTable</a></li><li data-type='method'><a href="SyncApi.html#resetSchema">resetSchema</a></li></ul></li><li><a href="Table.html">Table</a><ul class='methods'><li data-type='method'><a href="Table.html#convertColumnToMultilanguage">convertColumnToMultilanguage</a></li><li data-type='method'><a href="Table.html#convertColumnToSinglelanguage">convertColumnToSinglelanguage</a></li><li data-type='method'><a href="Table.html#createColumn">createColumn</a></li><li data-type='method'><a href="Table.html#createColumns">createColumns</a></li><li data-type='method'><a href="Table.html#createRow">createRow</a></li><li data-type='method'><a href="Table.html#createRowByObj">createRowByObj</a></li><li data-type='method'><a href="Table.html#createRows">createRows</a></li><li data-type='method'><a href="Table.html#deleteColumn">deleteColumn</a></li><li data-type='method'><a href="Table.html#fetch">fetch</a></li><li data-type='method'><a href="Table.html#findColumn">findColumn</a></li><li data-type='method'><a href="Table.html#getRow">getRow</a></li><li data-type='method'><a href="Table.html#getRows">getRows</a></li></ul></li><li><a href="TableBuilder.html">TableBuilder</a><ul class='methods'><li data-type='method'><a href="TableBuilder.html#create">create</a></li><li data-type='method'><a href="TableBuilder.html#displayName">displayName</a></li><li data-type='method'><a href="TableBuilder.html#group">group</a></li><li data-type='method'><a href="TableBuilder.html#hidden">hidden</a></li></ul></li><li><a href="Tables.html">Tables</a><ul class='methods'><li data-type='method'><a href="Tables.html#fetch">fetch</a></li><li data-type='method'><a href="Tables.html#find">find</a></li></ul></li></ul><h3>Global</h3><ul><li><a href="global.html#grudStructorizer">grudStructorizer</a></li></ul>
-    
-</nav>
-
-<div id="main">
-    
-    <h1 class="page-title">AsyncApi.js</h1>
-    
-
-    
-
-
-
-    
-    <section>
-        <article>
-            <pre class="prettyprint source linenums"><code>import Api from "./Api.js";
-
-/**
- *
- */
-class AsyncApi extends Api {
-  /**
-   *
-   * @param baseUrl {string}
-   * @param options {ApiOptions}
-   */
-  // eslint-disable-next-line no-useless-constructor
-  constructor(baseUrl, options) {
-    super(baseUrl, options);
-  }
-
-  /**
-   *
-   * @param method {string}
-   * @param url {string}
-   * @param [json] {object}
-   * @param [nonce] {string}
-   */
-  async doCall(method, url, json, nonce) {
-    const fullUrl = nonce ? this.baseUrl + url + "?" + new URLSearchParams({ nonce }) : this.baseUrl + url;
-
-    const options = {
-      method: method,
-      headers: this._getRequestHeaders(),
-      body: json ? JSON.stringify(json) : undefined
-    };
-
-    const response = await fetch(fullUrl, options);
-
-    return response.json();
-  }
-}
-
-export default AsyncApi;
-</code></pre>
-        </article>
-    </section>
-
-
-
-
-    
-    
-</div>
-
-<br class="clear">
-
-<footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.6.11</a> using the <a href="https://github.com/clenemt/docdash">docdash</a> theme.
-</footer>
-
-<script>prettyPrint();</script>
-<script src="scripts/polyfill.js"></script>
-<script src="scripts/linenumber.js"></script>
-
-
-
-</body>
-</html>
